Fix green domains export, again
The green domains export started failing again a few weeks ago due to
running out of memory, which, after some discussion, we decided was a
good excuse to introduce a TTL for domains (the vast majority of
pre-2025 domains have not been revisited since we started tracking cache
hits). In addition, we decided we needed to improve visibility of
failing cron jobs.

This commit fixes both these things:
 - it extends the clear_archived_domains command to also  delete
   greendomains which are over a year old, and renames it to
   clear_expired_domains

 - it introduces sentry monitoring for both the clear_archived_domains
   and dump_green_domains commands, used in the dataset export cronjobs.
